+++ tests/serialFunctionTests/serialDeviceTest.py
@@ -1,21 +1,13 @@
 # This function test is only for verifying serial devices
 
 from piGantry.piSerial import serialInter as serialComm
-import serial
 import time
+import csv
 
 
-micro = serialComm.serialObject(9600, "COM8")
-laser = serial.Serial(port= "COM16", baudrate = 9600, bytesize=serial.EIGHTBITS, timeout=10, write_timeout=10, parity=serial.PARITY_NONE, stopbits=serial.STOPBITS_ONE, rtscts=False)
-
-
-serialComm.initializeLaser(laser)
-
+micro = serialComm.serialObject(9600, "COM21")
 while True:
-    laserRead = serialComm.readLaser(laser)        
-    print(laserRead)
-    print(serialComm.readDial(micro.port))
-    time.sleep(0.1)
+    input = serialComm.readDial(micro.port)
+    print(input)
 
-