+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Changed
+- runZero now compares w/ float instead of messily cutting string
+- readDial now returns clean float instead of string
+- Refactored positional & keyword arguments & their references of runMoves, runZero, runOneMove to support dual motor & to be polymorphic
+- Changed references to X & Y to be agnostic (i.e 1 & 2 instead of X & Y) | This is so that order of motors is not hard coded & a separate function isn't required
+- Further abstracted runMoves function by taking out individual move & refactoring into new runOneMove function
 - Refactored zeroing to have pi feed steps directly to clearCore
 - Changed all references in changelog of unit tests to function tests
 - readDial flushes i/o before taking a reading
@@ -18,6 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - laser rangefinder functions in serialInter module now refactored for new laser rangefinder
 
 ### Removed
+- TkInter.py, redundant now that pyGUI module is ready
 - data_file.json, data_file1.json & data_file2.json. Outdated clutter
 - piSide.py, redundant now that serialInter module serialInterTest.py & ccSide.ino function test is ready
 - readSerialInput.ino, redundant now that serialInter module serialInterTest.py & ccSide.ino function test is ready
@@ -25,6 +31,17 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Outdated code in serialDeviceTest.py function test
 
 ### Added
+- Added motorFunc calcAvailableSPM function to handle generating motion profile
+- Added MM to Steps functionality to mathFunc calcDist
+- Middling UI
+- Motion parameter UI
+- pyGUI module for handling UI
+- ccSideY function test for 2nd clearCore
+- Non-micrometer support for runZero
+- Removed references of laser rangefinder except in serialInter module
+- Added second clearCore to serialInterTest function test
+- serialDeviceTest new framework for testing laser reading
+- Added new argument for laser functions to support continuous or one time reading
 - runZero function from serialInterTest function test now a function under motorFunc module
 - runMoves function from serialInterTest function test now a function under motorFunc module
 - initializeLaser function in serialInter module, handles sending start msg / addr to laser rangefinder